fix: use commnon style for Recommendation._base_criterion serialization
Before this change, the special style in which the criterion includes
not only its attributes but also its class in the serialization data
dictionary was when serializing the Recommendation._base_criterion
attribute.

The ActivePatients criterion was compatible with that style but other
criteria were not. As a result using any criterion but ActivePatients
as the value of the _base_criterion attribute led to (de)serialization
problems.