feat: population left dependent toggle
until now, the populations AND-combined with each of the intervention criteria.
as of now, we use "LeftDependentToggle" instead, to create "NOT_APPLICABLE" intervals
where population is not satisfied. this is required in order to be able to determine
how many simultaneous criteria can be present at most (in order to implement CappedMinCount)