fix: in find_rectangles, fix event sorting and adjacent interval merging
The comparison of events with the same timestamp and the same "track"
was too simplistic: for an open-and-close event pair for
single (point-shaped) interval, the open event should go before the
close event. however, for the close event from one interval and the
open event from a different, adjacent interval, the close event should
go before the open interval.

With the new sorting in place, the processing of event clusters can
restrict the "1 second lookahead" to open events (instead of all
events). The mechanism is intended to detect adjacent intervals by
possibly "pulling in" the open event of a following interval if that
open event happens in the next second after the close event of the
previous interval.