fix: add tie-breaker in compare_events in find_rectangles
The local function compare_events is used to sort events and thus
should induce a total order. Before this change, the final else
clauses of the function which handled the "don't care" case as

  else:
    return 1

did not induce a total order since it could lead to

      compare_events(a, b) == 1
  and compare_events(b, a) == 1
  => a < b and b < a

for certain events a, b.

This change introduces a tie-breaker criterion so that either (a < b
and not b < a) or the other way around.